requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
python-docx==1.1.0
openpyxl==3.1.2
flashtext==2.7
PyYAML==6.0.1
jinja2==3.1.4
pytest==8.3.2
httpx==0.27.2
